Clean up of options
